+++ src/app/(home)/page.tsx
@@ -1,4 +1,3 @@
-import PageViewTracker from '@/app/(private)/study-rooms/[id]/note/[noteId]/page-view-tracker';
 import {
   Session1,
   Session2,
@@ -11,6 +10,7 @@ import { homeMetadata } from '@/features/home/metadata';
 import { StructuredData } from '@/features/home/structured-data';
 import FloatingCTA from '@/features/home/ui/floating-cta';
 import { Footer } from '@/layout/footer';
+import { PageViewTracker } from '@/shared/components/gtm';
 import { cn } from '@/shared/lib';
 
 export const metadata = homeMetadata;
