+++ src/app/(private)/study-rooms/[id]/note/[noteId]/page.tsx
@@ -2,8 +2,7 @@ import BackLink from '@/features/dashboard/studynote/components/back-link';
 import StudyNoteDetailContentsSection from '@/features/dashboard/studynote/detail/components/contents-section';
 import StudyNoteDetailMetaSection from '@/features/dashboard/studynote/detail/components/meta-section';
 import { ColumnLayout } from '@/layout/column-layout';
-
-import PageViewTracker from './page-view-tracker';
+import { PageViewTracker } from '@/shared/components/gtm';
 
 type Props = {
   params: Promise<{ id: string; noteId: string }>;
