+++ src/shared/components/gtm/index.ts
@@ -1,0 +1,1 @@
+export { PageViewTracker } from './page-view-tracker';
